Fix compilation
Former-commit-id: b403cb57ad998c140fafa8db258b9e16a70bd295

+++ plugin/serializer/include/minko/file/AbstractWriter.hpp
@@ -36,7 +36,7 @@ namespace minko
 			typedef std::shared_ptr<AbstractWriter> Ptr;
 
 		private:
-			typedef std::vector<msgpack::type::tuple<short, short, std::string>> SerializedDependency;
+			typedef std::vector<msgpack::type::tuple<unsigned int, short, std::string>> SerializedDependency;
 
 		protected :
 			std::shared_ptr<Signal<Ptr>>	_complete;
